refactor: differentiate shared components and style
