+++ nus-faster-pay/src/Screens/AccountDashboard/Components/RightArrow.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 import { PaymentItem, PaymentCardSize } from "../../../Utils/Types";
-import { Subtitle, Body } from "../../../Utils/SharedComponents";
+import { Subtitle, Body } from "../../../Utils/StyledComponents";
 import { COLORS } from "../../../Utils/Colors";
 
 export function RightArrow() {
